requirements.txt
pandas
boto3
pyarrow
openai
python-dotenv
tiktoken
tqdm
